MoireSuperlattices

Documentation for MoireSuperlattices.

MoireSuperlattices.BLTMD

MoireSuperlattices.BLTMD

MoireSuperlattices.CommensurateBilayerHoneycomb

MoireSuperlattices.MoireReciprocalLattice

MoireSuperlattices.MoireSpace

MoireSuperlattices.MoireSpinor

MoireSuperlattices.MoireSystem

MoireSuperlattices.chemicalpotential

MoireSuperlattices.tribonds

MoireSuperlattices.trihoppings

MoireSuperlattices.trivalues

MoireSuperlattices.BLTMD — Type

BLTMD{T<:Number, L<:MoireReciprocalLattice, D<:Function, H<:OperatorGenerator} <: MoireSystem{H}

Twisted transition metal dichalcogenide homobilayers.

MoireSuperlattices.BLTMD — Method

BLTMD(a₀::Number, m::Number, θ::Number, Vᶻ::Number, μ::Number, V::Number, ψ::Number, w::Number; truncation::Int=4)

Continuum model of twisted transition metal dichalcogenide homobilayers.

Here, the parameters are as follows:

a₀: monolayer lattice constant (Å)

m: effective mass of the conduction band (mₑ)

θ: twist angle (°)

Vᶻ: perpendicular displacement field (meV)

μ: chemical potential (meV)

V: amplitude of Moire potential (meV)

ψ: phase of Moire potential (°)

w: interlayer hopping amplitude (meV)

MoireSuperlattices.CommensurateBilayerHoneycomb — Type

CommensurateBilayerHoneycomb

Commensurate Moire superlattice composed of two layers of honeycomb lattices.

MoireSuperlattices.MoireReciprocalLattice — Type

MoireReciprocalLattice{T<:Number} <: AbstractLattice{2, T, 0}

Moire reciprocal lattice with truncation.

MoireSuperlattices.MoireSpace — Type

MoireSpace <: SimpleInternal{MoireSpinor{Int, Int, Int, Rational{Int}, Int}}

The internal degrees of freedom of Moire systems.

MoireSuperlattices.MoireSpinor — Type

MoireSpinor{V<:Union{Int, Colon}, L<:Union{Int, Colon}, S<:Union{Int, Colon}, P<:Union{Rational{Int}, Colon}, N<:Union{Int, Colon}} <: SimpleIID

The index of the internal degrees of freedom of Moire systems.

MoireSuperlattices.MoireSystem — Type

MoireSystem{H<:OperatorGenerator} <: AbstractTBA{Fermionic{:TBA}, H, Nothing}

The continuum model of Moire systems.

MoireSuperlattices.chemicalpotential — Method

chemicalpotential(bltmd::Union{BLTMD, Algorithm{<:BLTMD}}, brillouinzone::BrillouinZone) -> Term

Get the chemical potential term.

MoireSuperlattices.tribonds — Function

tribonds(bltmd::BLTMD, truncation::Int, datatype=Float64) -> Vector{Vector{Bond}}

Get the tri-symmetric bonds on the triangular lattice.

MoireSuperlattices.trihoppings — Method

trihoppings(bltmd::Union{BLTMD, Algorithm{<:BLTMD}}, brillouin::BrillouinZone, truncation::Int; modulate::Bool=true) -> Tuple{Vararg{Term}}

Construct the three-fold symmetric hoppings of the highest energy level of a twisted TMD homobilayer up to the truncationth nearest neighbor on the triangular lattice.

MoireSuperlattices.trivalues — Method

trivalues(bltmd::BLTMD, neighbors::Vector{<:Vector{<:Bond}}, brillouin::BrillouinZone) -> Vector{Vector{ComplexF64}}

Get the tri-symmetric hopping amplitudes on the triangular lattice.