MoireSuperlattices.BLTMD — Type

BLTMD{
    L<:MoireReciprocalLattice,
    D<:Function,
    S<:OperatorGenerator,
    Q<:Quadraticization,
    H<:CategorizedGenerator{<:OperatorSum{<:Quadratic}}
} <: MoireSystem{NamedTuple{(:a₀, :m, :θ, :Vᶻ, :μ), NTuple{5, Float64}}, L, D, S, Q, H}

Twisted transition metal dichalcogenide homobilayers.

MoireSuperlattices.BLTMD — Method

BLTMD(a₀::Number, m::Number, θ::Number, Vᶻ::Number, μ::Number, V::Number, ψ::Number, w::Number; truncation::Int=4)

Continuum model of twisted transition metal dichalcogenide homobilayers.

Here, the parameters are as follows:

a₀: monolayer lattice constant (Å)

m: effective mass of the conduction band (mₑ)

θ: twist angle (°)

Vᶻ: perpendicular displacement field (meV)

μ: chemical potential (meV)

V: amplitude of Moire potential (meV)

ψ: phase of Moire potential (°)

w: interlayer hopping amplitude (meV)

MoireSuperlattices.CommensurateBilayerHoneycomb — Type

CommensurateBilayerHoneycomb

Commensurate Moire superlattice composed of two layers of honeycomb lattices.

MoireSuperlattices.MoireReciprocalLattice — Type

MoireReciprocalLattice{T<:Number} <: AbstractLattice{2, T, 0}

Moire reciprocal lattice with truncation.

MoireSuperlattices.MoireSpace — Type

MoireSpace <: SimpleInternal{MoireSpinor{Int, Int, Int, Rational{Int}, Int}}

The internal degrees of freedom of Moire systems.

MoireSuperlattices.MoireSpinor — Type

MoireSpinor{V<:Union{Int, Colon}, L<:Union{Int, Colon}, S<:Union{Int, Colon}, P<:Union{Rational{Int}, Colon}, N<:Union{Int, Colon}} <: SimpleInternalIndex

The index of the internal degrees of freedom of Moire systems.

MoireSuperlattices.MoireSuperlattice — Type

MoireSuperlattice{D<:Number} <: AbstractLattice{2, D, 2}

Abstract type of the emergent superlattices in Moire systems.

MoireSuperlattices.MoireSystem — Type

MoireSystem{P<:Parameters, L<:MoireReciprocalLattice, D<:Function, S<:OperatorGenerator, Q<:Quadraticization, H<:CategorizedGenerator{<:OperatorSum{<:Quadratic}}} <: TBA{Fermionic{:TBA}, H, Nothing}

The continuum model of Moire systems.

MoireSuperlattices.MoireTriangular — Type

MoireTriangular{N, D<:Number} <: MoireSuperlattice{D}

Emergent triangular superlattice in Moire systems.

MoireSuperlattices.MoireTriangular — Method

MoireTriangular(truncation::Int, vectors::AbstractVector{<:AbstractVector{<:Number}}; name=:MoireTriangular, origin=nothing)

Construct the emergent triangular superlattice in Moire systems.

QuantumLattices.Frameworks.Algorithm — Method

Algorithm(name::Symbol, bltmd::BLTMD, parameters::Parameters; kwargs...)

Construct an Algorithm with a BLTMD as the frontend.

QuantumLattices.Spatials.Lattice — Method

Lattice(moire::CommensurateBilayerHoneycomb, type::Symbol)

Get the minimum unit of the top/bottom layer of a commensurate Moire superlattice composed of two layers of honeycomb lattices.

Base.angle — Method

angle(moire::CommensurateBilayerHoneycomb) -> Float64

Get the twist angle of a commensurate Moire superlattice composed of two layers of honeycomb lattices.

Base.count — Method

count(moire::CommensurateBilayerHoneycomb) -> Int

Count the number of honeycomb unitcells contained in the unitcell of a commensurate Moire superlattice composed of two layers of honeycomb lattices.

The total number of atoms in the unitcell of the Moire superlattice is 4 times this result because of the AB sublattice and the top/bottom layer degrees of freedom.

MoireSuperlattices.coefficients — Method

coefficients(bltmd::BLTMD, lattice::MoireTriangular, brillouinzone::BrillouinZone; band::Int=dimension(bltmd)) -> Tuple{Vector{Vector{ComplexF64}}, Float64}
coefficients(bltmd::Algorithm{<:BLTMD}, lattice::MoireTriangular, brillouinzone::BrillouinZone; band::Int=dimension(bltmd.frontend)) -> Tuple{Vector{Vector{ComplexF64}}, Float64}

Get the coefficients of the hoppings and chemical potential of a bilayer TMD on the emergent triangular lattice.

MoireSuperlattices.terms — Method

terms(bltmd::BLTMD, lattice::MoireTriangular, brillouinzone::BrillouinZone; band::Int=dimension(bltmd), ismodulatable::Bool=true, tol=tol) -> NTuple{2*truncation(lattice)+1, Term}
terms(bltmd::Algorithm{<:BLTMD}, lattice::MoireTriangular, brillouinzone::BrillouinZone; band::Int=dimension(bltmd.frontend), ismodulatable::Bool=true, tol=tol) -> NTuple{2*truncation(lattice)+1, Term}

Get the hopping terms and chemical potential of a bilayer TMD on the emergent triangular lattice.

MoireSuperlattices.vectors — Method

vectors(moire::CommensurateBilayerHoneycomb) -> SVector{2, SVector{2, Float64}}

Get the translation vectors of a commensurate Moire superlattice composed of two layers of honeycomb lattices.

RecipesBase.apply_recipe — Function

plot(moire::CommensurateBilayerHoneycomb, choice::Symbol, n=2*ceil(Int, √count(moire)); topcolor=:red, bottomcolor=:blue, vector=true, vectorcolor=:green, moirecolor=:black, anglecolor=:grey)

Plot a Moire superlattice composed of two layers of honeycomb lattices in the real space or in the reciprocal space.